server_codes/requirements.txt
elasticsearch
paho-mqtt
numpy
matplotlib